+++ gittest/signing_fixtures.go
@@ -1,5 +1,4 @@
-// Package testsigning loads the shared GPG, SSH, and S/MIME test keys.
-package testsigning
+package gittest
 
 import (
 	"bytes"
@@ -101,5 +100,5 @@ func read(t testing.TB, name string) []byte {
 
 func fixturePath(name string) string {
 	_, here, _, _ := runtime.Caller(0)
-	return filepath.Join(filepath.Dir(here), "..", "..", "tests", "testdata", "signing", name)
+	return filepath.Join(filepath.Dir(here), "..", "tests", "testdata", "signing", name)
 }
