+++ tests/sign_gitea_verify_test.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/grafana/nanogit"
 	"github.com/grafana/nanogit/gittest"
-	"github.com/grafana/nanogit/internal/testsigning"
 	"github.com/grafana/nanogit/options"
 )
 
@@ -32,7 +31,7 @@ func TestSignGiteaVerify_GPG(t *testing.T) {
 	user.Token, err = server.CreateToken(ctx, user.Username)
 	require.NoError(t, err)
 
-	gpg := testsigning.LoadGPG(t)
+	gpg := gittest.LoadGPG(t)
 	const signerEmail = "signer@test.invalid"
 	setUserPrimaryEmail(t, server.URL(), user, signerEmail)
 	uploadGPGKey(t, server.URL(), user.Token, gpg.ArmoredPublic)
